[spec fix] Add scope to api routes
By adding a scope to the ideas route, and requiring our API end-users to
make use of that endpoint, we create some flexibility.
If ever we want to change our api, we keep this controller for the old
behavior, but we can add a new scope within the `path: :api` scope that
handles requests for another version of our api. This allows us to
depricate the api in a less rigid way.